requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Psycopg2==2.9.2
requests==2.26.0
NeedForCryptography~=1.5
SQLAlchemy~=1.4.27
APScheduler~=3.8.1
dataset~=1.5.1
python-dotenv~=0.19.2
waitress~=2.0.0
ecdsa~=0.17.0
pycryptodome==3.10.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
